requirements.txt
streamlit
ezdxf
matplotlib
plotly
numpy
numpy-stl
